add ViewModelTestBase to init mocks and set main disptacher for test
